REF: Add pylint and use recommendations
