+++ docs/conf.py
@@ -1,5 +1,4 @@
 #!/usr/bin/env python
-# -*- coding: utf-8 -*-
 #
 # rioxarray documentation build configuration file, created by
 # sphinx-quickstart on Fri Jun  9 13:47:02 2017.
