+++ rioxarray/__init__.py
@@ -1,11 +1,10 @@
-# -*- coding: utf-8 -*-
-
 """Top-level package for rioxarray."""
 
 __author__ = """rioxarray Contributors"""
 __email__ = "alansnow21@gmail.com"
 
-import rioxarray.rioxarray  # noqa
+import rioxarray.raster_array  # noqa
+import rioxarray.raster_dataset  # noqa
 from rioxarray._show_versions import show_versions  # noqa
 from rioxarray._version import __version__  # noqa
