+++ rioxarray/exceptions.py
@@ -1,4 +1,3 @@
-# -*- coding: utf-8 -*-
 """
 This contains exceptions for rioxarray.
 """
