+++ setup.py
@@ -1,4 +1,3 @@
-# -*- coding: utf-8 -*-
 """The setup script."""
 
 import os
