+++ test/conftest.py
@@ -2,7 +2,7 @@ import os
 
 from numpy.testing import assert_almost_equal, assert_array_equal
 
-from rioxarray.rioxarray import UNWANTED_RIO_ATTRS
+from rioxarray.raster_array import UNWANTED_RIO_ATTRS
 
 TEST_DATA_DIR = os.path.join(os.path.dirname(__file__), "test_data")
 TEST_INPUT_DATA_DIR = os.path.join(TEST_DATA_DIR, "input")
